pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "news-static"
version = "0.1.0"
description = "WeChat daily news static JSON generator"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.0",
  "Pillow>=10.4.0",
  "playwright>=1.52.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "PyYAML>=6.0.2",
  "selectolax>=0.3.21",
  "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "respx>=0.21.1",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
